// packages/core/src/generated/yarn_spinner.ts
export interface Program {
    /**
     * The name of the program.
     *
     * @generated from protobuf field: string name = 1;
     */
    name: string;
    /**
     * The collection of nodes in this program.
     *
     * @generated from protobuf field: map<string, Yarn.Node> nodes = 2;
     */
    nodes: {
        [key: string]: Node;
    };
    /**
     * The collection of initial values for variables; if a PUSH_VARIABLE
     * instruction is run, and the value is not found in the storage, this
     * value will be used
     *
     * @generated from protobuf field: map<string, Yarn.Operand> initial_values = 3;
     */
    initialValues: {
        [key: string]: Operand;
    };
}
/**
 * A collection of instructions
 *
 * @generated from protobuf message Yarn.Node
 */
export interface Node {
    /**
     * The name of this node.
     *
     * @generated from protobuf field: string name = 1;
     */
    name: string;
    /**
     * The list of instructions in this node.
     *
     * @generated from protobuf field: repeated Yarn.Instruction instructions = 7;
     */
    instructions: Instruction[];
    /**
     * The headers present on this node.
     *
     * @generated from protobuf field: repeated Yarn.Header headers = 6;
     */
    headers: Header[];
}
/**
 * @generated from protobuf message Yarn.Header
 */
export interface Header {
    /**
     * The name of the header.
     *
     * @generated from protobuf field: string key = 1;
     */
    key: string;
    /**
     * The value of the header.
     *
     * @generated from protobuf field: string value = 2;
     */
    value: string;
}
// ...
export interface Operand {
    /**
     * @generated from protobuf oneof: value
     */
    value: {
        oneofKind: "stringValue";
        /**
         * A string.
         *
         * @generated from protobuf field: string string_value = 1;
         */
        stringValue: string;
    } | {
        oneofKind: "boolValue";
        /**
         * A boolean (true or false).
         *
         * @generated from protobuf field: bool bool_value = 2;
         */
        boolValue: boolean;
    } | {
        oneofKind: "floatValue";
        /**
         * A floating point number.
         *
         * @generated from protobuf field: float float_value = 3;
         */
        floatValue: number;
    } | {
        oneofKind: undefined;
    };
}
// @generated message type with reflection information, may provide speed optimized methods
class Program$Type extends MessageType<Program> {
    constructor() {
        super("Yarn.Program", [
            { no: 1, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "nodes", kind: "map", K: 9 /*ScalarType.STRING*/, V: { kind: "message", T: () => Node } },
            { no: 3, name: "initial_values", kind: "map", K: 9 /*ScalarType.STRING*/, V: { kind: "message", T: () => Operand } }
        ]);
    }
    create(value?: PartialMessage<Program>): Program {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.name = "";
        message.nodes = {};
        message.initialValues = {};
        if (value !== undefined)
            reflectionMergePartial<Program>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Program): Program {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string name */ 1:
                    message.name = reader.string();
                    break;
                case /* map<string, Yarn.Node> nodes */ 2:
                    this.binaryReadMap2(message.nodes, reader, options);
                    break;
                case /* map<string, Yarn.Operand> initial_values */ 3:
                    this.binaryReadMap3(message.initialValues, reader, options);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    private binaryReadMap2(map: Program["nodes"], reader: IBinaryReader, options: BinaryReadOptions): void {
        let len = reader.uint32(), end = reader.pos + len, key: keyof Program["nodes"] | undefined, val: Program["nodes"][any] | undefined;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case 1:
                    key = reader.string();
                    break;
                case 2:
                    val = Node.internalBinaryRead(reader, reader.uint32(), options);
                    break;
                default: throw new globalThis.Error("unknown map entry field for field Yarn.Program.nodes");
            }
        }
        map[key ?? ""] = val ?? Node.create();
    }
    private binaryReadMap3(map: Program["initialValues"], reader: IBinaryReader, options: BinaryReadOptions): void {
        let len = reader.uint32(), end = reader.pos + len, key: keyof Program["initialValues"] | undefined, val: Program["initialValues"][any] | undefined;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case 1:
                    key = reader.string();
                    break;
                case 2:
                    val = Operand.internalBinaryRead(reader, reader.uint32(), options);
                    break;
                default: throw new globalThis.Error("unknown map entry field for field Yarn.Program.initial_values");
            }
        }
        map[key ?? ""] = val ?? Operand.create();
    }
    internalBinaryWrite(message: Program, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string name = 1; */
        if (message.name !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.name);
        /* map<string, Yarn.Node> nodes = 2; */
        for (let k of globalThis.Object.keys(message.nodes)) {
            writer.tag(2, WireType.LengthDelimited).fork().tag(1, WireType.LengthDelimited).string(k);
            writer.tag(2, WireType.LengthDelimited).fork();
            Node.internalBinaryWrite(message.nodes[k], writer, options);
            writer.join().join();
        }
        /* map<string, Yarn.Operand> initial_values = 3; */
        for (let k of globalThis.Object.keys(message.initialValues)) {
            writer.tag(3, WireType.LengthDelimited).fork().tag(1, WireType.LengthDelimited).string(k);
            writer.tag(2, WireType.LengthDelimited).fork();
            Operand.internalBinaryWrite(message.initialValues[k], writer, options);
            writer.join().join();
        }
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message Yarn.Program
 */
export const Program = new Program$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Node$Type extends MessageType<Node> {
    constructor() {
        super("Yarn.Node", [
            { no: 1, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 7, name: "instructions", kind: "message", repeat: 1 /*RepeatType.PACKED*/, T: () => Instruction },
            { no: 6, name: "headers", kind: "message", repeat: 1 /*RepeatType.PACKED*/, T: () => Header }
        ]);
    }
    create(value?: PartialMessage<Node>): Node {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.name = "";
        message.instructions = [];
        message.headers = [];
        if (value !== undefined)
            reflectionMergePartial<Node>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Node): Node {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string name */ 1:
                    message.name = reader.string();
                    break;
                case /* repeated Yarn.Instruction instructions */ 7:
                    message.instructions.push(Instruction.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                case /* repeated Yarn.Header headers */ 6:
                    message.headers.push(Header.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Node, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string name = 1; */
        if (message.name !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.name);
        /* repeated Yarn.Instruction instructions = 7; */
        for (let i = 0; i < message.instructions.length; i++)
            Instruction.internalBinaryWrite(message.instructions[i], writer.tag(7, WireType.LengthDelimited).fork(), options).join();
        /* repeated Yarn.Header headers = 6; */
        for (let i = 0; i < message.headers.length; i++)
            Header.internalBinaryWrite(message.headers[i], writer.tag(6, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message Yarn.Node
 */
export const Node = new Node$Type();
// ...
export const Header = new Header$Type();
// ...
export const Instruction = new Instruction$Type();
// ...
export const Operand = new Operand$Type();